backend/app/requirements.txt
# FastAPI后端服务依赖
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.5
aiofiles>=23.0.0

# 数据处理和验证
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0

# 日志和监控
structlog>=23.0.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0

# 数据库和缓存
sqlalchemy>=2.0.0
alembic>=1.12.0
redis>=4.5.0
asyncpg>=0.28.0

# WebRTC和实时通信
aiortc>=1.6.0
websockets>=11.0.0

# 安全和认证
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5

# 工具库
httpx>=0.24.0
pydantic-settings>=2.0.0
